+++ lib/openstudio-workflow.rb
@@ -36,6 +36,7 @@ require 'openstudio/workflow/version'
 require 'openstudio/workflow/multi_delegator'
 require 'openstudio/workflow/run'
 require 'openstudio/workflow/jobs/lib/apply_measures'
+require 'openstudio/workflow/time_logger'
 
 begin
   require 'openstudio'