+++ lib/openstudio/workflow/jobs/run_energyplus/run_energyplus.rb
@@ -21,7 +21,7 @@ class RunEnergyplus
   # Initialize
   # param directory: base directory where the simulation files are prepared
   # param logger: logger object in which to write log messages
-  def initialize(directory, logger, adapter, options = {})
+  def initialize(directory, logger, time_logger, adapter, options = {})
     energyplus_path = nil
     if /cygwin|mswin|mingw|bccwin|wince|emx/ =~ RUBY_PLATFORM
       energyplus_path = 'C:/EnergyPlus-8-1-0'
@@ -39,6 +39,7 @@ class RunEnergyplus
     @run_directory = "#{@directory}/run"
     @adapter = adapter
     @logger = logger
+    @time_logger = time_logger
     @results = {}
 
     @logger.info "#{self.class} passed the following options #{@options}"
@@ -98,6 +99,7 @@ class RunEnergyplus
 
     # can't create symlinks because the /vagrant mount is actually a windows mount
     @logger.info "Copying EnergyPlus files to run directory: #{@run_directory}"
+    @time_logger.start("Copying EnergyPlus files")
     FileUtils.copy("#{@options[:energyplus_path]}/libbcvtb.so", "#{@run_directory}/libbcvtb.so")
     FileUtils.copy("#{@options[:energyplus_path]}/libepexpat.so", "#{@run_directory}/libepexpat.so")
     FileUtils.copy("#{@options[:energyplus_path]}/libepfmiimport.so", "#{@run_directory}/libepfmiimport.so")
@@ -106,8 +108,11 @@ class RunEnergyplus
     FileUtils.copy("#{@options[:energyplus_path]}/ExpandObjects", "#{@run_directory}/ExpandObjects")
     FileUtils.copy("#{@options[:energyplus_path]}/EnergyPlus", "#{@run_directory}/EnergyPlus")
     FileUtils.copy("#{@options[:energyplus_path]}/Energy+.idd", "#{@run_directory}/Energy+.idd")
+    @time_logger.stop("Copying EnergyPlus files")
 
+    @time_logger.start("Running EnergyPlus")
     @results = call_energyplus
+    @time_logger.stop("Running EnergyPlus")
 
     @results
   end