+++ lib/openstudio/workflow/jobs/run_xml/run_xml.rb
@@ -22,7 +22,7 @@ require 'libxml'
 # This actually belongs as another class that gets added as a state dynamically
 class RunXml
   # RunXml
-  def initialize(directory, logger, adapter, options = {})
+  def initialize(directory, logger, time_logger, adapter, options = {})
     defaults = { use_monthly_reports: false, analysis_root_path: '.', xml_library_file: 'xml_runner.rb' }
     @options = defaults.merge(options)
     @directory = directory
@@ -31,6 +31,7 @@ class RunXml
     @adapter = adapter
     @results = {}
     @logger = logger
+    @time_logger = time_logger
     @logger.info "#{self.class} passed the following options #{@options}"
 
     # initialize instance variables that are needed in the perform section